Added ability to specify multiple element types within a station. Added ability to specify element orientations when using numerical element pattern data. Added option to specify metres or wavelength units for length of analytic dipole. Added option to normalise amplitude of all station beams at phase centre to 1.0. Added isotropic station beam option. Added run log and settings file to Measurement Set HISTORY table. Added settings file contents to visibility structure. Added functions to write out individual station configuration files to telescope model. Updated interfaces to various functions to accommodate necessary features. Updated ChangeLog.

+++ imaging/src/oskar_image_write.c
@@ -51,7 +51,7 @@ void oskar_image_write(const oskar_Image* image, oskar_Log* log,
     unsigned char grp = OSKAR_TAG_GROUP_IMAGE;
     FILE* stream;
     char* log_data = 0;
-    long log_size = 0;
+    size_t log_size = 0;
     const char* settings_path_str;
     const oskar_Mem* settings_path;
